Established parameters that are passed from page to page.
The FeelingResults that the user inputted are now passed to the next pages to be used for storing in the database and calculating the mindfulness activity and workout.
The text of the WorkoutOptionsPopup has also been changed to make it clearer to the user what the popup means.
Additional implementation is still needed regarding userID and database storage and retrieval to make everything work as intended.