Added the Signup page.
The signup page has been created. It functions identically to the login page, by which it doesn't actually perform its intended function and instead just redirects the user.
The Forgot Username and Forgot Password buttons, along with the new Create Account button, were all given a background color to emphasize the fact that they are buttons.