Added login and signup functionality.
The login and signup pages, aside from the Forgot Username and Forgot Password portions, should work as intended.
Supabase requires emails be verified, or an exception will be thrown when attempting to log in.
To temporarily circumvent this until we get the verification email system working, email verification in Supabase has been disabled.