Foundations for Forgot Password implemented.
Supabase's Reset Password email now has a link that leads to an index.html page. The purpose of this page is to redirect the user back to the app.
This index.html page is in another repository because I cannot access the Settings of this one to turn on GitHub Pages.
Here is the link: https://github.com/DevlinDelegard/ground_and_go_redirect
The index.html successfully brings the user back into the app without it crashing.
Now, a Forgot Password page must be created so that the user can be redirected to that instead of the login page.