Finalized the Forgot Password functionality.
Once a user clicks the link in the Supabase email and is redirected to the app, access and refresh tokens follow them back.
Using these tokens, the user is verified and is directed to the ForgotPasswordPage, where they can enter a new password.
After entering a new password, the auth.user Supabase table is updated with their new password.